Implement comparison
We remove the file format sniffing again
because apparently all the videos that end
in .ogg are only redirection pages now,
and trying to get the raw media content
results in 404 erorrs anyway, so the
likelihood that we download a big ogg
file that turns out to be a video
is very low.

This problem was the whole reason for
why format sniffing was thought to be
needed, now that it turns out to not
be a problem at all we can
remove sniffing.